// pong/src/utils/Types.ts
export interface DepthLineBehavior {
    movement?: 'vertical' | 'horizontal' | string;
    direction?: 'upwards' | 'downwards' | 'left' | 'right' | string;
    fade?: 'in' | 'out' | 'none' | string;
    
	// Pyramid specific properties
    pyramidBaseHeight?: number;
    pyramidBaseWidth?: number;
    pyramidPeakHeight?: number;
    pyramidPeakOffset?: number;

	// Ruin specicif properties
	ruinHSegments?: number;
	ruinTSegments?: number;
	maxHeight?: number;
	maxWidth?: number;
	segmentWidths?: number[];
    segmentHeights?: number[];
	hOffset?: number;

	// Wave specific properties
    waveAmplitude?: number;
    waveFrequency?: number;
    waveOffset?: number;
    phaseShift?: number;
    waveType?: 'sine' | 'square' | 'triangle' | 'sawtooth';
    amplitudeVariation?: number;   // Add this property
    minBottomHeight?: number | null;  // Add this property
}
// ...
export interface DepthLineOptions {
    initialized?: boolean;
    initialY?: number;
    velocityX?: number;
    velocityY?: number;
    width?: number;
    height?: number;
    upperLimit?: number;
    lowerLimit?: number;
    alpha?: number;
    alphaDecay?: number;
    alphaIncrease?: number;
    lifetime?: number;
    type?: string;
    despawn?: string;
    behavior?: DepthLineBehavior;
}

export interface PyramidDepthLineOptions extends DepthLineOptions {
    baseHeight?: number;
    peakHeight?: number;
    peakOffset?: number;
}

export interface RuinDepthLineOptions extends DepthLineOptions {
	ruinHSegments?: number;
	ruinTSegments?: number;
	maxHeight?: number;
	maxWidth?: number;
	segmentWidths?: number[];
    segmentHeights?: number[];
	hOffset?: number;
}

export interface WaveDepthLineOptions extends DepthLineOptions {
	amplitude?: number;
    frequency?: number; 
    waveOffset?: number;
    phaseShift?: number;
    resolution?: number;
    waveType?: 'sine' | 'square' | 'triangle' | 'sawtooth';
    amplitudeVariation?: number;   // Add this property
    minBottomHeight?: number | null;  // Add this property
}